Use the failure cache track of failing re-validation jobs.
Instead of storing the error count in the HttpCacheEntry, now the failure cache
is used. This avoids to hook into the complicated CachingHttpClient workflow
which takes different paths depending on the response and the set header flags.
Also we avoid the problem that the HttpCacheEntry might be flushed under
certain circumstances and thus the error code is lost.

Conflicts:

	httpclient-cache/src/test/java/org/apache/http/impl/client/cache/TestAsynchronousValidationRequest.java
	httpclient-cache/src/test/java/org/apache/http/impl/client/cache/TestAsynchronousValidator.java